Fix memoizing of usage statistics calls
Avoid minutes, seconds and milliseconds when calling usage statistics;
so that function call can be cached.

Enable a period task to cache usage statistics calls.
